Calculate clonality of all TCR data files in a given directory
This script calculates clonality of all TCR ( T cell receptor) sequencing files in a given directory and outputs a CSV file with the name of each file and corresponding clonality calculation.

Each file in the directory is TCR sequencing data of one a sample.
One column is the DNA sequences and the second column is the counts for each sequence.